#613 fix.

+++ Source/AliveLibAO/Midi.cpp
@@ -48,7 +48,7 @@ EXPORT __int16 CC SND_SsIsEos_DeInlined_477930(unsigned __int16 /*idx*/)
     return 0;
 }
 
-EXPORT int CC SFX_SfxDefinition_Play_477330(const SfxDefinition* /*sfxDef*/, signed int /*volLeft*/, signed int /*volRight*/, int /*pitch_min*/, signed __int16 /*pitch_max*/)
+EXPORT int CC SFX_SfxDefinition_Play_477330(const SfxDefinition* /*sfxDef*/, short /*volLeft*/, short /*volRight*/, short /*pitch_min*/, signed __int16 /*pitch_max*/)
 {
     NOT_IMPLEMENTED();
     return 0;
